+++ Web/RDD.Web.Tests/Serialization/PropertySerializerTests.cs
@@ -1,5 +1,4 @@
 ﻿using Rdd.Domain.Helpers.Expressions;
-using Rdd.Domain.Mocks;
 using Rdd.Domain.Models;
 using Rdd.Web.Tests.Models;
 using System.Collections.Generic;
@@ -25,18 +24,18 @@ namespace Rdd.Web.Tests.Serialization
         [Fact]
         public async Task HierarchySerialization()
         {
-            var entity = new Super();
-            var entity2 = new SuperSuper();
-            var fields = new ExpressionParser().ParseTree<Super>("");
+            var entity = new Domain.Tests.Models.Super();
+            var entity2 = new Domain.Tests.Models.SuperSuper();
+            var fields = new ExpressionParser().ParseTree<Domain.Tests.Models.Super>("");
 
             var json = await SerializeAsync(entity, fields);
             Assert.Contains("superProperty", json);
 
-            json = await SerializeAsync(new List<Hierarchy> { entity, entity2 }, fields);
+            json = await SerializeAsync(new List<Domain.Tests.Models.Hierarchy> { entity, entity2 }, fields);
             Assert.Contains("superProperty", json);
             Assert.Contains("superSuperProperty", json);
 
-            json = await SerializeAsync(new Selection<Hierarchy>(new List<Hierarchy> { entity, entity2 }, 0), fields);
+            json = await SerializeAsync(new Selection<Domain.Tests.Models.Hierarchy>(new List<Domain.Tests.Models.Hierarchy> { entity, entity2 }, 0), fields);
             Assert.Contains("superProperty", json);
             Assert.Contains("superSuperProperty", json);
         }