+++ Web/RDD.Web.Tests/ServerMock/Startup.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Rdd.Domain.Rights;
 using Rdd.Web.Helpers;
 
 namespace Rdd.Web.Tests.ServerMock
@@ -22,11 +23,15 @@ namespace Rdd.Web.Tests.ServerMock
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddRddSerialization<CurPrincipal>();
-
             services.AddDbContext<ExchangeRateDbContext>((service, options) => { options.UseInMemoryDatabase(databaseName: "Add_writes_to_database"); });
 
-            services.AddRdd<ExchangeRateDbContext, CombinationsHolder, CurPrincipal>();
+            services
+                .AddRdd<ExchangeRateDbContext>(rdd =>
+                {
+                    rdd.PagingLimit = 10;
+                    rdd.PagingMaximumLimit = 4242;
+                })
+                .WithDefaultRights(RightDefaultMode.Open);
 
             services.AddScoped<ExchangeRateController>();